+++ src/config/minimizers/css.d.ts
@@ -1,5 +1,5 @@
 import type { LiteralUnion } from '../../utility-types'
-import type { CSSNanoOptions } from '../npm'
+import type { CSSNanoOptions } from '../packages'
 import type { ConfigurablePlugin } from '../common'
 
 export type CSSMinimizer = LiteralUnion<'csso' | 'esbuild' | 'parcelCss'>
