+++ src/config/minimizers/index.d.ts
