+++ src/config/packages/esbuild-loader.d.ts
