+++ src/config/packages/html-webpack-plugin.d.ts
