+++ src/config/packages/mini-css-extract-plugin.d.ts
