+++ src/config/packages/swc-core.d.ts
