+++ src/config/platforms/index.d.ts
